Don't try to unlink a group.
Now that a command can point to a group in addition to generated files, we only want to try to remove the generated files during unlink_outputs.

+++ src/tup/updater.c
@@ -1598,14 +1598,16 @@ static int unlink_outputs(int dfd, struct node *n)
 	struct node *output;
 	LIST_FOREACH(e, &n->edges, list) {
 		output = e->dest;
-		if(unlinkat(dfd, output->tent->name.s, 0) < 0) {
-			if(errno != ENOENT) {
-				pthread_mutex_lock(&display_mutex);
-				show_result(n->tent, 1, NULL, NULL);
-				perror("unlinkat");
-				fprintf(stderr, "tup error: Unable to unlink previous output file: %s\n", output->tent->name.s);
-				pthread_mutex_unlock(&display_mutex);
-				return -1;
+		if(output->tent->type != TUP_NODE_GROUP) {
+			if(unlinkat(dfd, output->tent->name.s, 0) < 0) {
+				if(errno != ENOENT) {
+					pthread_mutex_lock(&display_mutex);
+					show_result(n->tent, 1, NULL, NULL);
+					perror("unlinkat");
+					fprintf(stderr, "tup error: Unable to unlink previous output file: %s\n", output->tent->name.s);
+					pthread_mutex_unlock(&display_mutex);
+					return -1;
+				}
 			}
 		}
 	}
